Rename filterlist to filterlists
