+++ intel/filterlists/keys.go
@@ -1,15 +1,15 @@
-package filterlist
+package filterlists
 
 const (
 	cacheDBPrefix = "cache:intel/filterlists"
 
 	// filterListCacheVersionKey is used to store the highest version
-	// of a filterlist file (base, intermediate or urgent) in the
+	// of a filterlists file (base, intermediate or urgent) in the
 	// cache database. It's used to decide if the cache database and
 	// bloomfilters need to be resetted and rebuilt.
 	filterListCacheVersionKey = cacheDBPrefix + "/version"
 
-	// filterListIndexKey is used to store the filterlist index.
+	// filterListIndexKey is used to store the filterlists index.
 	filterListIndexKey = cacheDBPrefix + "/index"
 
 	// filterListKeyPrefix is the prefix inside that cache database
