+++ intel/filterlists/lookup_map.go
@@ -1,4 +1,4 @@
-package filterlist
+package filterlists
 
 import "strings"
