+++ intel/filterlists/module_test.go
@@ -1,4 +1,4 @@
-package filterlist
+package filterlists
 
 /*
 func TestMain(m *testing.M) {
